Enumerate input/graphics devices with udev
We can query udev to get a list of devices more reliably and simply.
For example, the previous code does not detect controllers unless the
joydev module (old API) is loaded, whereas udev has already labeled the
correct event*, input*, and js* devices with ID_INPUT_JOYSTICK.

The same approach can probably be used for video/media devices, but I
didn't feel as confident with how to make that change.

The added pyudev dependency is documented in the README.