Add `--` to desktop entries to separate bubblejail's options
This makes sure that the original's desktop entry command line
options do not get interpreted as bubblejail options. This
does not happen right now because the instance arguments have a
`argparse.REMAINDER` number of arguments and all options following
instance name will get treated as arguments to run. However,
in the future number of arguments might be switched to `*` and
without `--` will look for options in the passed arguments.