Add object path filtering to D-Bus proxy
Apparently a lot of D-Bus services expose too many interfaces
and objects that are not required which increases attack surface.

For example, GNOME StatusNotifier D-Bus service allows calling
`/org/gnome/Shell.Eval` which executes javascript code.

So instead use the `--call` option to limit the object paths
that are allowed to be called.