Add "debug" service
This service allows passing raw arguments to bwrap and xdg-dbus-proxy.
It can be used both to debug bubblejail and as a crutch if bubblejail
missing any functionality.

Also it is not displayed in GUI because I don't want it to become
too much of a crutch. If there is somekind of common functionality
it should be separated in a service.