Rework CI and add lint script
Add tools/run_linters.py script to run the linters on the source
code. This makes it easier to integrate with CI.

Remove CodeQL from the CI because it does not do much and there are
much better native tools.

Add `ci.yaml` file with the new CI. Only run the pyflakes and mypy
for now. Use the Arch Linux container image.