Add support for alternative desktop entries names in profiles
For example, Chromium uses either `chromium-browser.desktop` or
`chromium.desktop` depending on a distro.

Add support for defining desktop entry name as a list. When some
code wants to get a path to the desktop entry iterate over the
list and return the first file that exists. If none exist return
None.