Read info fd until end of file
It is possible to the bwrap to write only a fragment of JSON.
In that case the JSON decode will fail and bubblejail will become
stuck.

The bwrap will close the passed file descriptor once it is done
writing the JSON. Use this to read the info fd until the EOF and
only decode it after end is reached.

The EOF is reached once reads return empty.