Add `pasta_network` service
This is alternative network stack. It seems to be more actively
developed than `slirp4netns` and support advanced features such
as port forwarding or device binding.

The service only has one setting `extra_args` which takes a list
of strings to be added to the command line. See `passt` man page
for available options.